==> notebooks/9_private_board.ipynb <==
import pandas as pd
import numpy as np
import tensorflow as tf
from sklearn.cross_validation import train_test_split
import xgboost as xgb
from scipy import sparse
from sklearn.feature_extraction import FeatureHasher
from scipy.sparse import coo_matrix,csr_matrix,csc_matrix, hstack
from sklearn.preprocessing import normalize
from sklearn.utils import shuffle
from sklearn import linear_model
import gc
from sklearn import preprocessing

import matplotlib.pyplot as plt
%matplotlib inline

predictors_target_11 = ['ruta_freq', 'clien_freq', 'agen_freq', 'prod_freq',
       'agen_for_log_de', 'ruta_for_log_de', 'cliente_for_log_de',
       'producto_for_log_de', 'agen_ruta_for_log_de',
       'agen_cliente_for_log_de', 'agen_producto_for_log_de',
       'ruta_cliente_for_log_de', 'ruta_producto_for_log_de',
       'cliente_producto_for_log_de', 'cliente_for_log_sum',
       'cliente_producto_agen_for_log_sum', 'corr', 't_min_6', 't_min_2',
       't_min_3', 't_min_4', 't_min_5', 't2_min_t6', 't3_min_t6',
       't4_min_t6', 't5_min_t6', 't2_min_t3', 't2_min_t4', 't2_min_t5',
       't3_min_t4', 't3_min_t5', 't4_min_t5', 'LR_prod', 'LR_prod_corr',
       'target', 't_m_6_cum', 't_m_5_cum', 't_m_4_cum', 't_m_3_cum',
       't_m_2_cum', 'NombreCliente', 'weight', 'weight_per_piece', 'pieces']

predictors_11 = ['ruta_freq', 'clien_freq', 'agen_freq',
       'prod_freq', 'agen_for_log_de', 'ruta_for_log_de',
       'cliente_for_log_de', 'producto_for_log_de', 'agen_ruta_for_log_de',
       'agen_cliente_for_log_de', 'agen_producto_for_log_de',
       'ruta_cliente_for_log_de', 'ruta_producto_for_log_de',
       'cliente_producto_for_log_de', 'cliente_for_log_sum',
       'cliente_producto_agen_for_log_sum', 'corr', 't_min_6', 't_min_2',
       't_min_3', 't_min_4', 't_min_5', 't2_min_t6', 't3_min_t6',
       't4_min_t6', 't5_min_t6', 't2_min_t3', 't2_min_t4', 't2_min_t5',
       't3_min_t4', 't3_min_t5', 't4_min_t5', 'LR_prod', 'LR_prod_corr',
       't_m_6_cum', 't_m_5_cum', 't_m_4_cum', 't_m_3_cum', 't_m_2_cum',
       'NombreCliente', 'weight', 'weight_per_piece', 'pieces']

train_pivot_xgb_time2 = pd.read_csv('train_pivot_34567_to_9.csv',usecols = predictors_target_11,dtype = np.float32)

train_pivot_xgb_time2.columns.values

train_pivot_xgb_time2.head()

train_pivot_xgb_time2.shape

train_pivot_56789_to_11 = pd.read_csv('train_pivot_56789_to_11_private.csv',dtype=np.float32)

train_pivot_56789_to_11.head()

train_pivot_56789_to_11.columns.values

### begin xgboost
---------------

train_feature_11 = train_pivot_xgb_time2.drop(['target'],axis = 1)
train_label_11 = train_pivot_xgb_time2['target']

dtrain_11 = xgb.DMatrix(train_feature_11,label = train_label_11,missing= np.nan)

#right now the best is this parameter set with 700 round
param_11 = {'booster':'gbtree',
         'nthread': 8,
         'max_depth':8, 
         'eta':0.015,
         'silent':1,
         'subsample':0.5, 
         'objective':'reg:linear',
         'eval_metric':'rmse',
         'colsample_bytree':0.5}

# # for cv
# num_round = 1400

# cvresult = xgb.cv(param_11, dtrain_11, num_round, nfold=4,show_stdv=False,
#                         seed = 0, verbose_eval = 1)
# print(cvresult.tail())

# for real train
num_round = 2200
evallist = [(dtrain_11,'train')]
bst = xgb.train(param_11, dtrain_11, num_round,evallist)

- 1600 : 0.442933
- 1800 : 0.442292
- 2200 : 0.441234

bst.save_model('bst11_2200_eta0015.model')

dtest_11 = xgb.DMatrix(train_pivot_56789_to_11.drop(['id'],axis =1), missing=np.nan)

submission_11_all_train = pd.DataFrame()
submission_11_all_train = train_pivot_56789_to_11[['id']].copy()
submission_11_all_train['predict'] = bst.predict(dtest_11)
submission_11_all_train.reset_index(drop = True,inplace = True)

submission_11_all_train['predict'].describe()

submission_11_all_train.head()

submission_11_all_train['id'] = submission_11_all_train['id'].astype(int)

### concat result

submission_10 = pd.read_csv('week10_private.csv',index_col = 0)

submission_10.head()

submission_all_train = pd.concat([submission_10,submission_11_all_train],axis =0)
submission_all_train['predict'] = submission_all_train['predict'].apply(np.expm1)
submission_all_train.rename(columns = {'predict':'Demanda_uni_equil'},inplace = True)


submission_all_train['Demanda_uni_equil'] = submission_all_train['Demanda_uni_equil'].round(1)
submission_all_train.loc[submission_all_train['Demanda_uni_equil']<0,'Demanda_uni_equil'] = 0
submission_all_train['Demanda_uni_equil'].describe()

submission_all_train['id'] = submission_all_train['id'].astype(int)

submission_all_train.head()

submission_all_train.to('')

### validation 45678 to 10
------------------------

train_pivot_45678_to_10 = pd.read_pickle('validation_45678_10.pickle')
train_pivot_45678_to_10.head()

dtest_valid = xgb.DMatrix(train_pivot_45678_to_10.drop(['id'],axis =1), missing=np.nan)

submission_10_valid = pd.DataFrame()
submission_10_valid = train_pivot_45678_to_10[['id']].copy()
submission_10_valid['predict'] = bst.predict(dtest_valid)
submission_10_valid.reset_index(drop = True,inplace = True)

submission_10_valid.head()

submission_all_train = pd.DataFrame()
submission_all_train = pd.concat([submission_10_valid,submission_11_all_train],axis =0)
submission_all_train['predict'] = submission_all_train['predict'].apply(np.expm1)
submission_all_train.rename(columns = {'predict':'Demanda_uni_equil'},inplace = True)
submission_all_train.head()

submission_all_train.loc[submission_all_train['Demanda_uni_equil']<0,'Demanda_uni_equil'] = 0
submission_all_train['Demanda_uni_equil'].describe()

submission_all_train['id'] = submission_all_train['id'].astype(int)

submission_all_train['Demanda_uni_equil'].describe()

submission_all_train.head()

submission_all_train['Demanda_uni_equil'] = submission_all_train['Demanda_uni_equil'].round(1)

submission_all_train.to_csv('valid10_6.csv',index = False)
